I started Learning Design Patterns There are Three types of Design Pattern 1. Creational 2. Structural 3. Behavioral
I Learnt Factory Design Pattern (Creational Design Pattern) today
Understood the Importance and what was problem with the
naive Solution of creating objects of derived classes
and
How this Solution Beautifully and easily solve the Problem of Creating objects.

This Design Pattern is used whenever we have to create on object of
many types of derived classes.